- README: update formatting and cleanup
- Fix undefined placeholder warnings
- take picture: red error message
- Allow admins to choose what gets deleted at reset (inspired by https://github.com/andreknieriem/photobooth/issues/178)
  - always:
    - delete db.txt
  - optional (but enabled by default):
    - delete images
    - delete "mail-addresses.txt
    - delete personal config (my.config.inc.php)
- Add possibility to choose dark countdown background (inspired by https://github.com/andreknieriem/photobooth/issues/198 )
- Allow defining Photobooth web server IP to fix image download via QR-Code if Photobooth is accessed via localhost/127.0.0.1

+++ api/admin.php
@@ -10,21 +10,32 @@ if (!isset($data['type'])) {
 }
 
 if ($data['type'] == 'reset') {
-    // empty folders
-    foreach ($config['foldersAbs'] as $folder) {
-        if (is_dir($folder)) {
-            $files = glob($folder.'/*.jpg');
-            foreach ($files as $file) { // iterate files
-                if (is_file($file)) {
-                    unlink($file); // delete file
+    if($config['reset_remove_images']) {
+        // empty folders
+        foreach ($config['foldersAbs'] as $folder) {
+            if (is_dir($folder)) {
+                $files = glob($folder.'/*.jpg');
+                foreach ($files as $file) { // iterate files
+                    if (is_file($file)) {
+                        unlink($file); // delete file
+                    }
                 }
             }
         }
     }
 
-    // delete mail-addresses.txt
-    if(is_file('../mail-addresses.txt')){
-        unlink('../mail-addresses.txt');
+    if($config['reset_remove_mailtxt']) {
+        // delete mail-addresses.txt
+        if(is_file('../mail-addresses.txt')){
+            unlink('../mail-addresses.txt');
+        }
+    }
+
+    if($config['reset_remove_config']) {
+        // delete personal config
+        if(is_file('../config/my.config.inc.php')){
+            unlink('../config/my.config.inc.php');
+        }
     }
 
     // delete db.txt
